+++ API Project/src/main/java/com/bounswe/Application.java
@@ -1,4 +1,4 @@
-package hello;
+package com.bounswe;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
