+++ API Project/src/main/java/com/bounswe/controllers/GreetingController.java
@@ -1,10 +1,11 @@
-package hello;
+package com.bounswe.controllers;
 
 import java.util.concurrent.atomic.AtomicLong;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 import java.util.ArrayList;
+import com.bounswe.models.Greeting;
 
 @RestController
 public class GreetingController {
