+++ API Project/src/main/java/com/bounswe/controllers/ItemController.java
@@ -1,4 +1,4 @@
-package hello;
+package com.bounswe.controllers;
 
 import java.util.concurrent.atomic.AtomicLong;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -9,6 +9,9 @@ import sun.security.util.Length;
 
 import java.util.ArrayList;
 import java.util.Date;
+import com.bounswe.models.User;
+import com.bounswe.models.CulturalHeritageItem;
+
 @RestController
 public class ItemController{
