+++ API Project/src/main/java/com/bounswe/controllers/UsersController.java
@@ -1,9 +1,10 @@
-package hello;
+package com.bounswe.controllers;
 
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 import java.util.ArrayList;
+import com.bounswe.models.User;
 
 @RestController
 public class UsersController {
